Moved the ai training stuff outside of manifoldSaveData into aiPressureUtil
