+++ OASMan_ESP32/src/manifoldSaveData.h
@@ -89,25 +89,6 @@ public:
     Preferencable mlSampleRecord; // ML_SAMPLE_RECORD_VERSION (only version — weights are never persisted)
 };
 
-// One pressure-offset sample: the flowing sensor readings captured just before a valve closed, plus the
-// settled bag reading after it closed. The model learns offset = settled_bag - raw_bag. See AI_TRAINING.md.
-struct PressureLearnSaveStruct
-{
-    uint8_t raw_bag;     // flowing bag reading captured just before the valve closed
-    uint8_t settled_bag; // settled bag reading after close (label)
-    uint8_t raw_tank;    // flowing tank reading at capture
-    void print()
-    {
-        Serial.print("{");
-        Serial.print((int)raw_bag);
-        Serial.print(", ");
-        Serial.print((int)settled_bag);
-        Serial.print(", ");
-        Serial.print((int)raw_tank);
-        Serial.print("}");
-    }
-};
-
 struct ProfileRaw {
     int profileNum;
     byte pressure[4];
@@ -120,17 +101,6 @@ void requestSendConfigBT();
 void beginSaveData();
 ProfileRaw readProfile(byte profileIndex);
 void savePressuresToProfile(byte profileIndex, float _WHEEL_FRONT_PASSENGER, float _WHEEL_REAR_PASSENGER, float _WHEEL_FRONT_DRIVER, float _WHEEL_REAR_DRIVER);
-
-PressureLearnSaveStruct *getLearnData(SOLENOID_AI_INDEX aiIndex);
-int getLearnDataLength(SOLENOID_AI_INDEX aiIndex);
-
-void clearPressureData();
-
-// Append one offset sample to the model's SPIFFS file + RAM mirror (the training task re-fits from it).
-void recordLearnSample(SOLENOID_AI_INDEX aiIndex, uint8_t raw_bag, uint8_t settled_bag, uint8_t raw_tank);
-
-// The 4 offset models are RAM-only, re-fit from the stored samples every boot (nothing persisted).
-OffsetModel *getOffsetModel(SOLENOID_AI_INDEX aiIndex);
 
 headerDefineSaveFunc(riseOnStart, bool);
 headerDefineSaveFunc(maintainPressure, bool);
