+++ OASMan_ESP32/src/pressureMath.cpp
@@ -1,5 +1,5 @@
 #include "pressureMath.h"
-#include "manifoldSaveData.h" // PressureLearnSaveStruct (for refit)
+#include "aiPressureUtil.h" // PressureLearnSaveStruct (for refit)
 
 #define ML_N ML_NUM_COEFF
